inlined position directive
It was wrapping each position span with a span of the container; in
turn, this made the layout harder. It’s possible to have a directive
completely replace the containing element content… but that appears to
be necessary by using a link() method and doing it manually. At that
point, I have to ask, why bother having the directive in the first
place?