Move the initialization of the last frame drawn into link()
This is a consequence of the Directive life cycle. The directive is
created once per application lifetime, but link() is called once per
compilation/rendering (which, in this case, happens when the game board
is initialized/rendered).

Without this fix, there’s a bug in the rendering. Moving from a
partially-drawn hangman to a more-completely-drawn-hangman (i.e.: with
a higher target frame) caused only the subsequent frames to be drawn
(ex: I saw a pair of legs with no gallows or body). This was because
the canvas was being cleared but lastFrameDrawn still retained it’s
earlier state. The two were inconsistent with each other, and so the
redrawn canvas only showed the later frames.

With this change, lastFrameDrawn is always reset when the directive is
compiled and the canvas redrawn.